+++ gems/aws-sdk-core/lib/aws-sdk-core/assume_role_credentials.rb
@@ -54,6 +54,7 @@ module Aws
         end
       end
       @client = client_opts[:client] || STS::Client.new(client_opts)
+      @async_refresh = true
       super
     end