+++ gems/aws-sdk-core/lib/aws-sdk-core/instance_profile_credentials.rb
@@ -79,6 +79,7 @@ module Aws
       @token_ttl = options[:token_ttl] || 21_600
       @token = nil
       @no_refresh_until = nil
+      @async_refresh = false
       super
     end