+++ gems/aws-sdk-core/spec/aws/assume_role_credentials_spec.rb
@@ -91,6 +91,17 @@ module Aws
       expect(c.expiration).to eq(in_one_hour)
     end
 
+    it 'refreshes asynchronously' do
+      # expiration 6 minutes out, within the async exp time window
+      allow(credentials).to receive(:expiration).and_return(Time.now + (6*60))
+      expect(client).to receive(:assume_role).at_least(2).times
+      expect(Thread).to receive(:new).and_yield
+      c = AssumeRoleCredentials.new(
+        role_arn: 'arn',
+        role_session_name: 'session')
+      c.credentials
+    end
+
     it 'refreshes credentials automatically when they are near expiration' do
       allow(credentials).to receive(:expiration).and_return(Time.now)
       expect(client).to receive(:assume_role).exactly(4).times