+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Asynchronously refresh AWS credentials (#2641).
+
 * Issue - Add x-amz-region-set to list of headers deleted for re-sign.
 
 3.129.1 (2022-03-10)